+++ upcloud/service/retry_test.go
@@ -2,9 +2,11 @@ package service
 
 import (
 	"context"
+	"fmt"
 	"testing"
 	"time"
 
+	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -78,6 +80,54 @@ func TestRetry_noInterval(t *testing.T) {
 
 			assert.Error(t, err)
 			assert.Nil(t, value)
+		})
+	}
+}
+
+func problemsThenValue(status, count int) func(i int, _ context.Context) (*string, error) {
+	return func(i int, _ context.Context) (*string, error) {
+		if i < count {
+			return nil, &upcloud.Problem{Status: status}
+		}
+
+		value := "ready"
+		return &value, nil
+	}
+}
+
+func TestRetry_retry500(t *testing.T) {
+	t.Parallel()
+
+	testcases := []struct {
+		status      int
+		count       int
+		expectValue bool
+		expectError bool
+	}{
+		{status: 500, count: 3, expectValue: true, expectError: false},
+		{status: 500, count: 4, expectValue: false, expectError: true},
+		{status: 499, count: 1, expectValue: false, expectError: true},
+	}
+
+	for _, test := range testcases {
+		name := fmt.Sprintf("HTTP %d x %d", test.status, test.count)
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+
+			ctx := context.TODO()
+			value, err := retry(ctx, problemsThenValue(test.status, test.count), &retryConfig{interval: time.Millisecond * 125})
+
+			if test.expectError {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
+
+			if test.expectValue {
+				assert.Equal(t, "ready", *value)
+			} else {
+				assert.Nil(t, value)
+			}
 		})
 	}
 }